# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Care Platform API",
        "description": "Complete API documentation for Care Platform - A comprehensive marketplace connecting service providers with customers",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Care Platform API Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/bookings": {
            "get": {
                "tags": [
                    "Admin - Bookings"
                ],
                "summary": "Get all bookings",
                "operationId": "e840c2aa99df300af325e05e26ca2656",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/bookings/statistics": {
            "get": {
                "tags": [
                    "Admin - Bookings"
                ],
                "summary": "Get booking statistics",
                "operationId": "9bd5efc06c70b0a7a549310d633123bf",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/bookings/{id}": {
            "get": {
                "tags": [
                    "Admin - Bookings"
                ],
                "summary": "Get booking details",
                "operationId": "ccfa4b39cd3e1304f980054bdfc1a6c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Bookings"
                ],
                "summary": "Delete booking permanently",
                "operationId": "3c8bbb476450bd6406c201d4e756455f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/bookings/{id}/cancel": {
            "put": {
                "tags": [
                    "Admin - Bookings"
                ],
                "summary": "Cancel booking",
                "operationId": "714116adf5623a8a479ab605c562143c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cancelled successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}